requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
Django==5.1.6
fonttools==4.56.0
frozenlist==1.5.0
future==1.0.0
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huepy==1.2.1
idna==3.8
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.0.0
instabot==0.117.0
jiter==0.8.2
kiwisolver==1.4.8
lxml==5.3.0
magic-filter==1.0.12
matplotlib==3.10.1
mock==5.1.0
moviepy==2.1.2
multidict==6.1.0
numpy==2.2.3
openai==0.28.0
packaging==24.2
pillow==10.4.0
pluggy==1.5.0
proglog==0.1.10
propcache==0.2.1
psutil==6.0.0
pydantic==2.10.5
pydantic_core==2.27.2
pyfiglet==1.0.2
pyparsing==3.2.1
PySocks==1.7.1
pystray==0.19.5
pyTelegramBotAPI==4.22.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
python-telegram-bot==21.10
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
responses==0.25.6
scapy==2.5.0
schedule==1.2.2
screeninfo==0.8.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
termcolor==2.4.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.2
XlsxWriter==3.2.0
yarl==1.18.3
